requirements/base.txt
pytz==2022.7.1  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.1  # https://github.com/redis/redis-py

# Django
# ------------------------------------------------------------------------------
django  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
pandas==2.0.1 # https://github.com/pandas-dev/pandas
celery==5.2.7
django-celery-beat==2.5.0
XlsxWriter==3.1.0
zipp==3.15.0
python-decouple==3.8